const correctness and rename/delete commands

+++ src/commands/implementations/DeleteCommand.cpp
@@ -6,7 +6,16 @@ namespace cpp2 {
             : AbstractCommand(clientConnection, syncManager) {}
 
     bool DeleteCommand::execute() {
-        auto relativePath = clientConnection.waitForIncomingMessage();
+        const auto relativePath = clientConnection.waitForIncomingMessage();
+
+        if (!fileSystemManager.pathExists(relativePath)) {
+            throw std::logic_error{ERROR_NO_SUCH_PATH};
+        }
+
+        if (!fileSystemManager.hasWritePermissions(relativePath)) {
+            throw std::logic_error{ERROR_NO_PERMISSION};
+        }
+
         fileSystemManager.deletePath(relativePath);
         clientConnection.sentOutgoingMessage(OK_RESPONSE);
         return true;
